tests/wpt/tests/html/canvas/offscreen/text/2d.text.measure.lang.inherit.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.text.measure.lang.inherit</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.text.measure.lang.inherit</h1>
<p class="desc">Testing the lang attribute</p>


<script>
promise_test(async t => {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  document.documentElement.setAttribute('lang','tr');
  var f = new FontFace("Lato-Medium", "url('/fonts/Lato-Medium.ttf')");
  f.load();
  document.fonts.add(f);
  await document.fonts.ready;

  ctx.font = '50px Lato-Medium';
  ctx.lang = 'inherit';
  const text = 'fi';
  const tm_tr = ctx.measureText(text);
  const tr_width = tm_tr.width;

  ctx.lang = 'en';
  const tm_en = ctx.measureText(text);
  const en_width = tm_en.width;

  _assert(tr_width > en_width, "tr_width > en_width");

}, "Testing the lang attribute");
</script>
